fix(ios): simplify wda proxy client
